Add module part for TemporaryFileStream.h
