+++ src/DataStreams/TemporaryFileStream.h
@@ -1,16 +1,10 @@
 #pragma once
 
-#include <DataStreams/IBlockInputStream.h>
-#include <DataStreams/NativeBlockInputStream.h>
-#include <DataStreams/NativeBlockOutputStream.h>
-#include <DataStreams/copyData.h>
-#include <Processors/QueryPipelineBuilder.h>
-#include <Processors/Executors/PullingPipelineExecutor.h>
 #include <Processors/ISource.h>
+#include <Processors/QueryPipelineBuilder.h>
 #include <Compression/CompressedReadBuffer.h>
-#include <Compression/CompressedWriteBuffer.h>
 #include <IO/ReadBufferFromFile.h>
-#include <IO/WriteBufferFromFile.h>
+#include <DataStreams/IBlockStream_fwd.h>
 
 namespace DB
 {
@@ -22,72 +16,28 @@ struct TemporaryFileStream
     CompressedReadBuffer compressed_in;
     BlockInputStreamPtr block_in;
 
-    explicit TemporaryFileStream(const std::string & path)
-        : file_in(path)
-        , compressed_in(file_in)
-        , block_in(std::make_shared<NativeBlockInputStream>(compressed_in, DBMS_TCP_PROTOCOL_VERSION))
-    {}
-
-    TemporaryFileStream(const std::string & path, const Block & header_)
-        : file_in(path)
-        , compressed_in(file_in)
-        , block_in(std::make_shared<NativeBlockInputStream>(compressed_in, header_, 0))
-    {}
+    explicit TemporaryFileStream(const std::string & path);
+    TemporaryFileStream(const std::string & path, const Block & header_);
 
     /// Flush data from input stream into file for future reading
-    static void write(const std::string & path, const Block & header, QueryPipelineBuilder builder, const std::string & codec)
-    {
-        WriteBufferFromFile file_buf(path);
-        CompressedWriteBuffer compressed_buf(file_buf, CompressionCodecFactory::instance().get(codec, {}));
-        NativeBlockOutputStream output(compressed_buf, 0, header);
-
-        auto pipeline = QueryPipelineBuilder::getPipeline(std::move(builder));
-        PullingPipelineExecutor executor(pipeline);
-
-        output.writePrefix();
-
-        Block block;
-        while (executor.pull(block))
-            output.write(block);
-
-        output.writeSuffix();
-        compressed_buf.finalize();
-    }
+    static void write(const std::string & path, const Block & header, QueryPipelineBuilder builder, const std::string & codec);
 };
+
 
 class TemporaryFileLazySource : public ISource
 {
 public:
-    TemporaryFileLazySource(const std::string & path_, const Block & header_)
-        : ISource(header_)
-        , path(path_)
-        , done(false)
-    {}
-
+    TemporaryFileLazySource(const std::string & path_, const Block & header_);
     String getName() const override { return "TemporaryFileLazySource"; }
 
 protected:
-    Chunk generate() override
-    {
-        if (done)
-            return {};
-
-        if (!stream)
-            stream = std::make_unique<TemporaryFileStream>(path, header);
-
-        auto block = stream->block_in->read();
-        if (!block)
-        {
-            done = true;
-            stream.reset();
-        }
-        return Chunk(block.getColumns(), block.rows());
-    }
+    Chunk generate() override;
 
 private:
     const std::string path;
     Block header;
     bool done;
+
     std::unique_ptr<TemporaryFileStream> stream;
 };
